+++ include/httpServer.h
@@ -5,11 +5,13 @@
 #include <memory>
 #include <functional>
 #include <thread>
+#include <atomic>
 
 #include "logger.h"
 #include "tcpSocket.h"
 #include "httpRequest.h"
 #include "httpResponse.h"
+#include "indexPage.h"
 
 namespace butterfly
 {
@@ -22,11 +24,11 @@ class HTTPServer
 
 private:
     unsigned int _port;
-    bool _running;
+    std::atomic<bool> _running;
     std::shared_ptr<TCPSocket> _tcpSocket, _newTCPSocket;
     std::unique_ptr<HTTPRequest> _httpRequest;
     std::unique_ptr<HTTPResponse> _httpResponse;
-    std::thread t1;
+    std::thread _serverThread;
     /**
      * Handles incoming requests to the http server
      *
@@ -65,7 +67,11 @@ private:
      */
     void errorResponse(size_t statusCode);
 
+    /**
+     * Private run method for the HTTP Server
+     */
     void _run();
+
 public:
 
     /**
@@ -73,7 +79,7 @@ public:
      *
      * Usage:
      *      std::shared_ptr<butterfly::HTTPServer> server = std::make_shared<butterfly::HTTPServer>(8080);
-     *      server.run();
+     *      server->run();
      *
      * @param port: Port for the Server
      */