+++ src/httpServer.cpp
@@ -6,6 +6,9 @@ namespace butterfly
 
 HTTPServer::HTTPServer(unsigned int port) : _port(port),  _running(false), _tcpSocket(std::make_shared<TCPSocket>()), _newTCPSocket(std::make_shared<TCPSocket>())
 {
+    #ifdef LOGGING
+    LOG_TRACE("Create class HTTPServer");
+    #endif
 
     _tcpSocket->bind(_port);
     _tcpSocket->listen();
@@ -13,7 +16,6 @@ HTTPServer::HTTPServer(unsigned int port) : _port(port),  _running(false), _tcpS
 
 HTTPServer::~HTTPServer()
 {
-    _tcpSocket->disconnect();
     stop();
 }
 
@@ -36,16 +38,23 @@ void HTTPServer::_run()
 void HTTPServer::run(bool blocking)
 {
     _running = true;
-    t1 = std::thread(&HTTPServer::_run, this);
+    _serverThread = std::thread(&HTTPServer::_run, this);
+    #ifdef LOGGING
+    LOG_INFO("Running HTTPServer on port " << _port);
+    #endif
 
     if (blocking)
     {
-        t1.join();
+        _serverThread.join();
+    } else
+    {
+        _serverThread.detach();
     }
 }
 
 void HTTPServer::stop()
 {
+    _tcpSocket->disconnect();
     _running = false;
 }
 
@@ -114,7 +123,7 @@ void HTTPServer::browserRoute()
     _httpResponse->setStatusCode(302);
     _httpResponse->setReasonPhrase(_httpResponse->getStatusCode());
     _httpResponse->setHTTPHeader("Content-Type", "text/html; charset=utf8");
-    _httpResponse->addBody("<!DOCTYPE html><html><body><h1>YOU HAVE BEEN HACKED!!!</h1></body></html>");
+    _httpResponse->addBody(butterfly::INDEXPAGE);
     _httpResponse->setHTTPHeader("Content-Length", std::to_string(_httpResponse->getBody().length()));
 
     _httpResponse->prepareOutgoing();