+++ test/crypto/aesDecryptorTest.cpp
@@ -46,7 +46,7 @@ TEST_F(AESDecryptorTest, decryptFile)
         }
     } else
     {
-        std::cerr << "Could not decrypt file " << TESTFILE + butterfly::params::ENC_BFLY_FILE_ENDING << " because file does not exists!" << std::endl;
+        std::cerr << "Could not decrypt file " << TESTFILE + butterfly::params::ENC_BFLY_FILE_ENDING << " because file does not exist!" << std::endl;
     }
 
 }