+++ CHANGELOG.md
@@ -2,15 +2,27 @@
 All notable changes to project [butterfly](https://github.com/bierschi/butterfly) will be documented in this file
 
 
-## [1.2] -
+## [1.3] -
 
 ### Added
 
-- TorSocket integration
-- Support of different rsa key sizes
-- Support of different aes key sizes
-- Integration of HTTP Server
-- Open browser with a payment page after encryption
+- Added butterfly image
+- Integration of TorSocket class
+- Enhancement of the browser payment page after encryption
+- Prevention of Out-of-Memory errors with huge files
+- Improved CI/CD Pipeline
+- Integration of docker containers
+- Added cppcheck to repository
+- Added toolchain for multiple platforms
+
+## [1.2] - 2022-06-02
+
+### Added
+
+- Support of different RSA key sizes
+- Support of different AES key sizes
+- Integration of an HTTP Server/Client
+- Open browser with a basic html page after encryption
 
 ## [1.1] - 2022-02-11
 
@@ -21,11 +33,12 @@ All notable changes to project [butterfly](https://github.com/bierschi/butterfly
 - Webserver bflyServerApp for a secure remote decryption
 - Notes section for encryption/decryption with the openssl cli
 - Tools section for common actions
-- Improved argument parser
+- Improvements of the argument parser
 
 ## [1.0] - 2021-09-17
 
 ### Added
+
 - First runnable version with basic encryption/decryption capability
 
 ##  Development Start
@@ -38,6 +51,7 @@ All notable changes to project [butterfly](https://github.com/bierschi/butterfly
 - Hybrid Encryption/Decryption mechanism
 
 
+[1.3]: https://github.com/bierschi/butterfly/compare/v1.2...v1.3
 [1.2]: https://github.com/bierschi/butterfly/compare/v1.1...v1.2
 [1.1]: https://github.com/bierschi/butterfly/compare/v1.0...v1.1
 [1.0]: https://github.com/bierschi/butterfly/releases/tag/v1.0