+++ webserver/CHANGELOG.rst
@@ -1,2 +1,3 @@
 
 16.01.2022: Start of project bflyServerApp
+26.10.2022: Updated version to 1.1.0